LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/zero-intrinsic-iteration-duration.tentative-expected.txt


FAIL Intrinsic iteration duration is non-negative assert_equals: 'actual' unit type must be 'percent' for "Default duration is 100%" expected (string) "percent" but got (undefined) undefined
